Move rack middleware
This adds a central location of the existing rack middleware.

Also renames the old `RackMiddleware` to be more descriptive.
